some new change

+++ ERPBO/Agriculture/DomainModels/FinishGoodRecipeDetails.cs
@@ -15,7 +15,7 @@ namespace ERPBO.Agriculture.DomainModels
         public long FGRDetailsId { get; set; }
         public long RawMaterialId { get; set; }
         public string ReceipeBatchCode { get; set; }
-        public int FGRRawMaterQty { get; set; }
+        public double FGRRawMaterQty { get; set; }
         public string FGRRawMaterUnit { get; set; }
         public long OrganizationId { get; set; }
         public long? EUserId { get; set; }
